utils/calculate_parameters.py
from collections import OrderedDict
import PIL
from PIL import Image

def calculate_parameters(model):
    parameters = 0
    for weight in model.parameters():
        p = 1
        for dim in weight.size():
            p *= dim
        parameters += p
    return parameters