requirements.txt
accelerate==0.18.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colour-demosaicing==0.2.3
colour-science==0.4.2
contourpy==1.0.7
correlation==1.0.0
cycler==0.11.0
dataclasses==0.6
deformable-attention==0.0.16
einops==0.6.0
exceptiongroup==1.1.1
execnet==1.9.0
fastrlock==0.8.1
filelock==3.10.7
fonttools==4.39.3
h5py==3.8.0
huggingface-hub==0.13.3
idna==3.4
imageio==2.27.0
importlib-resources==5.12.0
iniconfig==2.0.0
joblib==1.2.0
jpeg4py==0.1.4
kiwisolver==1.4.4
lazy_loader==0.2
linformer==0.2.1
loguru==0.6.0
lpips==0.1.4
matplotlib==3.7.1
natsort==8.3.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.3
psutil==5.9.4
ptflops==0.6.9
pycparser==2.21
pyparsing==3.0.9
pytest==7.2.2
pytest-xdist==3.2.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tensorboardX==2.6
tifffile==2023.3.21
timm==0.6.13
tomli==2.0.1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.65.0
urllib3==1.26.15
yacs==0.1.8
zipp==3.15.0
